continuing to work on error handling and updates

+++ client/src/Categories/IconSelector.jsx
@@ -3,13 +3,13 @@ import { useState, useContext, useEffect } from "react"
 
 // internal dependencies
 import capitalize from "../Services/capitalize"
-import handleChanges from "../Services/handleChanges"
 import { statusContext } from "../Services/Context"
 import apiService from "../Services/apiService"
 
 // components
 import Button from "../Components/Button"
 import Statusbar from "../Components/Statusbar"
+import RegularField from "../Components/RegularField"
 
 //------ MODULE INFO
 // Displays a modal that allows the user to select a new icon for a category.
@@ -20,7 +20,8 @@ import Statusbar from "../Components/Statusbar"
 const IconSelector = ({ changes, setChanges, toggle }) => {
 
     // get context
-    const { status, setStatus } = useContext(statusContext)
+    const { setStatus } = useContext(statusContext)
+    const [ forceValidation, setForceValidation ] = useState(0)
 
     // fetch icons from the api
     // the newIcons state causes the call to be made again when new icons are uploaded.
@@ -99,6 +100,8 @@ const IconSelector = ({ changes, setChanges, toggle }) => {
             setPreview(null)
             return
         }
+
+        console.log(file)
         const filepath = URL.createObjectURL(file)
         setPreview(filepath)
 
@@ -108,59 +111,88 @@ const IconSelector = ({ changes, setChanges, toggle }) => {
 
     // handles icon upload
     const handleUpload = () => {
-        if (file) {
 
-            // validation
-            if (uploaderChanges.name === "") {
-                setStatus({
-                    message: "An icon label sets the hover and alt text for that icon, and helps people understand what it is. Please set one before uploading.",
-                    error: true
-                })
-                return
-            } else if (iconList.map(icon => icon.name).includes(uploaderChanges.name)) {
-                setStatus({
-                    message: "An icon with that name already exists. If you're sure you want to upload this icon, give it a different name.",
-                    error: true
-                })
-                return
-            }
-
-            // get the file extension and date
-            const ext = file.type.split("/")[1]
-            const date = new Date().getTime()
-
-            // create the icon object
-            const iconSubmission = {
-                name: uploaderChanges.name.toLowerCase(),
-                file,
-                date,
-                ext
-            }
-            
-            // send the icon object to the api
-            apiService.uploadIcon(iconSubmission, (res) => {
-                if (res.error) {
-                    setStatus({
-                        message: "We were not able to upload your icon.",
-                        error: true
-                    })
-                } else {
-                    setStatus({
-                        message: `The icon ${ res.name } has been created.`,
-                        error: false
-                    })
-                    setNewIcons(res.name)
-                    setUploadForm(false)
-                }
+        if (!file) {
+            setStatus({
+                message: "You have not selected a file for upload.",
+                error: true
             })
+            return
         }
+
+        // label validation
+        if (uploaderChanges.name.length > 255) {
+            setStatus({
+                message: "This icon label is too long.",
+                error: true
+            })
+            return
+        } else if (uploaderChanges.name === "") {
+            setStatus({
+                message: "An icon label sets the hover and alt text for that icon, and helps people understand what it is. Please set one before uploading.",
+                error: true
+            })
+            return
+        } else if (iconList.map(icon => icon.name).includes(uploaderChanges.name)) {
+            setStatus({
+                message: "An icon with that name already exists. If you're sure you want to upload this icon, give it a different name.",
+                error: true
+            })
+            return
+        }
+
+        // get the file extension and date
+        const ext = file.type.split("/")[1]
+        const date = new Date().getTime()
+
+        // validate the file itself
+        if (ext !== "jpg" && ext !== "jpeg" && ext !== "png") {
+            setStatus({
+                message: "Icons must be in JPG or PNG format.",
+                error: true
+            })
+            return
+        }
+        if (file.size > 1000000) {
+            setStatus({
+                message: "Icons must be smaller than 1 MB.",
+                error: true
+            })
+            return
+        }
+
+        // create the icon object
+        const iconSubmission = {
+            name: uploaderChanges.name.toLowerCase(),
+            file,
+            date,
+            ext
+        }
+        
+        // send the icon object to the api
+        apiService.uploadIcon(iconSubmission, (res) => {
+            if (res.error) {
+                setStatus({
+                    message: "We were not able to upload your icon.",
+                    error: true
+                })
+            } else {
+                setStatus({
+                    message: `The icon ${ res.name } has been created.`,
+                    error: false
+                })
+                setNewIcons(res.name)
+                setUploadForm(false)
+            }
+        })
     }
 
     // the unsaved changes window doesn't appear here, but our form handling still expects the state
     // likewise, a changes object is expected rather than a string
     const [ unsaved, setUnsaved ] = useState(false)
     const [ uploaderChanges, setUploaderChanges ] = useState({
-        name: ""
+        name: "",
+        errorFields: []
     })
 
     // toggles mass delete mode
@@ -211,6 +243,13 @@ const IconSelector = ({ changes, setChanges, toggle }) => {
     // sends the icons to delete as an array of id numbers to the api
     const confirmDelete = () => {
         const count = iconsToDelete.length
+        if (count > 1) {
+            setStatus({
+                message: "You have not selected any icons to delete.",
+                error: true
+            })
+            return
+        }
         apiService.deleteIcons(iconsToDelete, data => {
             if (data.error) {
                 setStatus({
@@ -250,6 +289,14 @@ const IconSelector = ({ changes, setChanges, toggle }) => {
         }
     }
 
+    const formControls = { 
+        changes: uploaderChanges,
+        setChanges: setUploaderChanges,
+        unsaved,
+        setUnsaved, 
+        force: forceValidation
+    }
+
     return (
         <div id="icon-selector-box">
             <div className="row title-row mt-3 mb-2">
@@ -282,7 +329,7 @@ const IconSelector = ({ changes, setChanges, toggle }) => {
             <div id="icon-uploader" className="row row-info">
                 <div className="col col-info">
                     <div className="col-head">
-                        File
+                        File *
                     </div>
                     <div className="col-content">
                         <input type="file" id="uploader" accept="image/png,image/jpg,image/jpeg" onChange={ (e) => setFile(e.target.files[0]) } />
@@ -299,14 +346,14 @@ const IconSelector = ({ changes, setChanges, toggle }) => {
                 </div>
                 <div className="col col-info">
                     <div className="col-head">
-                        Label
+                        Label *
                     </div>
                     <div className="col-content">
-                        <input className='my-2'
-                            type="text" 
-                            name="name" 
-                            value={ uploaderChanges.name } 
-                            onChange={ (event) => handleChanges.handleTextChange(event, uploaderChanges, setUploaderChanges, setUnsaved) } 
+                        <RegularField 
+                            type="text"
+                            name="name"
+                            formControls={ formControls }
+                            required={ true }
                         />
                     </div>
                 </div>
